refactor: commonify AddFolder and FolderSettings
