+++ frontend/components.d.ts
@@ -18,7 +18,6 @@ declare module 'vue' {
     AddContactPhoneModal: typeof import('./src/components/Modals/AddContactPhoneModal.vue')['default']
     AddDomainModal: typeof import('./src/components/Modals/AddDomainModal.vue')['default']
     AddEmailModal: typeof import('./src/components/Modals/AddEmailModal.vue')['default']
-    AddFolderModal: typeof import('./src/components/Modals/AddFolderModal.vue')['default']
     AddMailingListExternalMemberModal: typeof import('./src/components/Modals/AddMailingListExternalMemberModal.vue')['default']
     AddMailingListInternalMembersModal: typeof import('./src/components/Modals/AddMailingListInternalMembersModal.vue')['default']
     AddMailingListMemberInput: typeof import('./src/components/AddMailingListMemberInput.vue')['default']
@@ -50,7 +49,7 @@ declare module 'vue' {
     EmailContent: typeof import('./src/components/EmailContent.vue')['default']
     EmojiPicker: typeof import('./src/components/EmojiPicker.vue')['default']
     ExportSettings: typeof import('./src/components/Settings/ExportSettings.vue')['default']
-    FolderSettingsModal: typeof import('./src/components/Modals/FolderSettingsModal.vue')['default']
+    FolderModal: typeof import('./src/components/Modals/FolderModal.vue')['default']
     FrappeLogo: typeof import('./src/components/Icons/FrappeLogo.vue')['default']
     HeaderActions: typeof import('./src/components/HeaderActions.vue')['default']
     IdentitySettings: typeof import('./src/components/Settings/IdentitySettings.vue')['default']
